Re-fix #197 - ios-sim does not work for iPad-Pro for iOS 10 runtime
The name of the device is not identical in 'Devices' and 'Device Types'. In 'Devices' for example, the 'iPad Pro (9.7 inch)' is 'iPad Pro (9.7-inch)' (note the hyphen).
This makes iso-sim not match the device with the device type, which we need to get the runtime and device id support.