Remove process.exit when simctl prerequisite is not present.
The init() function will return a shell return code. In the CLI context, when the return code is non-zero, it will then call process.exit with that return code.

Since the behaviour has changed, the version is bumped a major so that dependents don't pick this up automatically and encounter unexpected results.